#61 changed one last function to arrow func + renaming file
