+++ src/pages/experiences.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Button } from '@mui/material';
 import { Download } from '@mui/icons-material';
 import PageHeading from '../components/PageHeading';
-import ResumeExperiencesSection from '../components/ResumePage/ResumeExperiencesSection';
+import ResumeExperiencesSection from '../components/ResumePage/ResumeExperiences';
 import { RESUME_GOOGLE_DRIVE_ID } from '../data/Pages/ResumePageData';
 
 const ExperiencePage = () => {
